Module aws_sdk_kinesisanalytics::error

Modules

code_validation_exception

See CodeValidationException

concurrent_modification_exception

See ConcurrentModificationException

invalid_application_configuration_exception

See InvalidApplicationConfigurationException

invalid_argument_exception

See InvalidArgumentException

limit_exceeded_exception

See LimitExceededException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

resource_provisioned_throughput_exceeded_exception

See ResourceProvisionedThroughputExceededException

service_unavailable_exception

See ServiceUnavailableException

too_many_tags_exception

See TooManyTagsException

unable_to_detect_schema_exception

See UnableToDetectSchemaException

unsupported_operation_exception

See UnsupportedOperationException

Structs

AddApplicationCloudWatchLoggingOptionError

AddApplicationInputError

AddApplicationInputProcessingConfigurationError

AddApplicationOutputError

AddApplicationReferenceDataSourceError

CodeValidationException

User-provided application code (query) is invalid. This can be a simple syntax error.

ConcurrentModificationException

Exception thrown as a result of concurrent modification to an application. For example, two individuals attempting to edit the same application at the same time.

CreateApplicationError

DeleteApplicationCloudWatchLoggingOptionError

DeleteApplicationError

DeleteApplicationInputProcessingConfigurationError

DeleteApplicationOutputError

DeleteApplicationReferenceDataSourceError

DescribeApplicationError

DiscoverInputSchemaError

InvalidApplicationConfigurationException

User-provided application configuration is not valid.

InvalidArgumentException

Specified input parameter value is invalid.

LimitExceededException

Exceeded the number of applications allowed.

ListApplicationsError

ListTagsForResourceError

ResourceInUseException

Application is not available for this operation.

ResourceNotFoundException

Specified application can't be found.

ResourceProvisionedThroughputExceededException

Discovery failed to get a record from the

ServiceUnavailableException

The service is unavailable. Back off and retry the operation.

StartApplicationError

StopApplicationError

TagResourceError

TooManyTagsException

Application created with too many tags, or too many tags added to an application. Note that the maximum number of application tags includes system tags. The maximum number of user-defined application tags is 50.

UnableToDetectSchemaException

Data format is not valid. Amazon Kinesis Analytics is not able to detect schema for

UnsupportedOperationException

The request was rejected because a specified parameter is not supported or a specified resource is not valid for this operation.

UntagResourceError

UpdateApplicationError

Enums

AddApplicationCloudWatchLoggingOptionErrorKind

AddApplicationInputErrorKind

AddApplicationInputProcessingConfigurationErrorKind

AddApplicationOutputErrorKind

AddApplicationReferenceDataSourceErrorKind

CreateApplicationErrorKind

DeleteApplicationCloudWatchLoggingOptionErrorKind

DeleteApplicationErrorKind

DeleteApplicationInputProcessingConfigurationErrorKind

DeleteApplicationOutputErrorKind

DeleteApplicationReferenceDataSourceErrorKind

DescribeApplicationErrorKind

DiscoverInputSchemaErrorKind

ListApplicationsErrorKind

ListTagsForResourceErrorKind

StartApplicationErrorKind

StopApplicationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApplicationErrorKind