Module aws_sdk_kinesisanalytics::model

Modules

application_detail

See ApplicationDetail

application_summary

See ApplicationSummary

application_update

See ApplicationUpdate

cloud_watch_logging_option

See CloudWatchLoggingOption

cloud_watch_logging_option_description

See CloudWatchLoggingOptionDescription

cloud_watch_logging_option_update

See CloudWatchLoggingOptionUpdate

csv_mapping_parameters

See CsvMappingParameters

destination_schema

See DestinationSchema

input

See Input

input_configuration

See InputConfiguration

input_description

See InputDescription

input_lambda_processor

See InputLambdaProcessor

input_lambda_processor_description

See InputLambdaProcessorDescription

input_lambda_processor_update

See InputLambdaProcessorUpdate

input_parallelism

See InputParallelism

input_parallelism_update

See InputParallelismUpdate

input_processing_configuration

See InputProcessingConfiguration

input_processing_configuration_description

See InputProcessingConfigurationDescription

input_processing_configuration_update

See InputProcessingConfigurationUpdate

input_schema_update

See InputSchemaUpdate

input_starting_position_configuration

See InputStartingPositionConfiguration

input_update

See InputUpdate

json_mapping_parameters

See JsonMappingParameters

kinesis_firehose_input

See KinesisFirehoseInput

kinesis_firehose_input_description

See KinesisFirehoseInputDescription

kinesis_firehose_input_update

See KinesisFirehoseInputUpdate

kinesis_firehose_output

See KinesisFirehoseOutput

kinesis_firehose_output_description

See KinesisFirehoseOutputDescription

kinesis_firehose_output_update

See KinesisFirehoseOutputUpdate

kinesis_streams_input

See KinesisStreamsInput

kinesis_streams_input_description

See KinesisStreamsInputDescription

kinesis_streams_input_update

See KinesisStreamsInputUpdate

kinesis_streams_output

See KinesisStreamsOutput

kinesis_streams_output_description

See KinesisStreamsOutputDescription

kinesis_streams_output_update

See KinesisStreamsOutputUpdate

lambda_output

See LambdaOutput

lambda_output_description

See LambdaOutputDescription

lambda_output_update

See LambdaOutputUpdate

mapping_parameters

See MappingParameters

output

See Output

output_description

See OutputDescription

output_update

See OutputUpdate

record_column

See RecordColumn

record_format

See RecordFormat

reference_data_source

See ReferenceDataSource

reference_data_source_description

See ReferenceDataSourceDescription

reference_data_source_update

See ReferenceDataSourceUpdate

s3_configuration

See S3Configuration

s3_reference_data_source

See S3ReferenceDataSource

s3_reference_data_source_description

See S3ReferenceDataSourceDescription

s3_reference_data_source_update

See S3ReferenceDataSourceUpdate

source_schema

See SourceSchema

tag

See Tag

Structs

ApplicationDetail

ApplicationSummary

ApplicationUpdate

Describes updates to apply to an existing Amazon Kinesis Analytics application.

CloudWatchLoggingOption

Provides a description of CloudWatch logging options, including the log stream

CloudWatchLoggingOptionDescription

Description of the CloudWatch logging option.

CloudWatchLoggingOptionUpdate

Describes CloudWatch logging option updates.

CsvMappingParameters

Provides additional mapping information when the record format

DestinationSchema

Describes the data format when records are written to the destination.

Input

When you configure the application input, you specify the streaming source, the

InputConfiguration

When you start your application,

InputDescription

Describes the application input configuration.

InputLambdaProcessor

An object that contains the Amazon Resource Name (ARN) of the AWS Lambda function that

InputLambdaProcessorDescription

An object that contains the Amazon Resource Name (ARN) of the AWS Lambda function that

InputLambdaProcessorUpdate

Represents an update to the InputLambdaProcessor that is used to preprocess the records in the stream.

InputParallelism

Describes the number of in-application streams to create for a given

InputParallelismUpdate

Provides updates to the parallelism count.

InputProcessingConfiguration

Provides a description of a processor that is used to preprocess the records in the

InputProcessingConfigurationDescription

Provides configuration information about an input processor. Currently, the only input processor available is

InputProcessingConfigurationUpdate

Describes updates to an InputProcessingConfiguration.

InputSchemaUpdate

Describes updates for the application's input schema.

InputStartingPositionConfiguration

Describes the point at which the application reads from

InputUpdate

Describes updates to a specific input configuration (identified by the

JsonMappingParameters

Provides additional mapping information when JSON is the record format on the streaming source.

KinesisFirehoseInput

Identifies an Amazon Kinesis Firehose delivery stream as the streaming source. You

KinesisFirehoseInputDescription

KinesisFirehoseInputUpdate

When updating application input configuration, provides information about an

KinesisFirehoseOutput

When configuring application output,

KinesisFirehoseOutputDescription

KinesisFirehoseOutputUpdate

KinesisStreamsInput

Identifies an Amazon Kinesis stream as the streaming source. You provide the

KinesisStreamsInputDescription

KinesisStreamsInputUpdate

When updating application input configuration, provides information about an

KinesisStreamsOutput

When configuring application output, identifies an Amazon Kinesis stream as the

KinesisStreamsOutputDescription

KinesisStreamsOutputUpdate

LambdaOutput

When configuring application output, identifies an AWS Lambda function as the

LambdaOutputDescription

For an application output, describes the AWS Lambda function configured as its

LambdaOutputUpdate

When updating an output configuration using the UpdateApplication

MappingParameters

When configuring application input

Output

OutputDescription

Describes the application output configuration, which includes

OutputUpdate

RecordColumn

Describes the mapping of each data element in the streaming

RecordFormat

ReferenceDataSource

Describes the reference data source by providing the source information (S3 bucket name and object key name), the resulting in-application table name that is created, and the necessary schema to map the data elements in the Amazon S3 object to the in-application table.

ReferenceDataSourceDescription

Describes the reference data source configured for an application.

ReferenceDataSourceUpdate

When you update a reference data source configuration for an application, this object provides all the updated values (such as the source bucket name and object key name), the in-application table name that is created, and updated mapping information that maps the data in the Amazon S3 object to the in-application reference table that is created.

S3Configuration

Provides a description of an Amazon S3 data source, including the Amazon Resource

S3ReferenceDataSource

Identifies the S3 bucket and object that contains the reference data. Also identifies the IAM role Amazon Kinesis Analytics can assume to read this object on your behalf.

S3ReferenceDataSourceDescription

Provides the bucket name and object key name that stores the reference data.

S3ReferenceDataSourceUpdate

Describes the S3 bucket name, object key name, and IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf and populate the in-application reference table.

SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

Tag

A key-value pair (the value is optional) that you can define and assign to AWS resources. If you specify a tag that already exists, the tag value is replaced with the value that you specify in the request. Note that

Enums

ApplicationStatus

InputStartingPosition

RecordFormatType